+++ ip_tests/ORB/test_orb.py
@@ -1,0 +1,8 @@
+import cv2
+
+__all__=["Image_ORB"]
+
+def Image_ORB(im):
+    orb=cv2.ORB_create()
+    kp,des=orb.detectAndCompute(im,None)
+    return (des,kp)
